requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7